+++ docs/fr_FR/Changelog.md
@@ -7,6 +7,8 @@
 -   Nouveauté: Nouvelle cmd 'act_colorChoice' pour remplacer 'setRGB'.
 -   Améliorations: Support Ikea INSPELNING (2781).
 -   Améliorations: Modèle cmde 'Colour' remplacé par 'act_colorChoice'.
+-   Améliorations: Suppression cmdes 'colorRouge/colorVert/colorBleu' => cmd 'act_ColorChoice'.
+-   Améliorations: Suppression cmdes 'setRouge/setVert/setBleu' => cmd 'act_ColorChoice'.
 
 ## 251228-BETA-1